--- users/seth/rules.mk
@@ -1,1 +1,0 @@
-SRC += seth.c
